tests/integration/test_list_user_organizations.py
#!/usr/bin/env python3
import json
import logging
import os
import sys
import time
import requests
import firebase_admin
from firebase_admin import firestore
from firebase_admin import auth
from datetime import datetime

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Path to the test organization ID file
TEST_ORG_ID_FILE = os.path.join(os.path.dirname(__file__), '../test_data/test_org_id.txt')

# Initialize Firebase Admin SDK
try:
    firebase_admin.get_app()
except ValueError:
    # Use the application default credentials
    firebase_admin.initialize_app()

# API endpoint for list_user_organizations
API_ENDPOINT = "https://europe-west3-relexro.cloudfunctions.net/relex-backend-list-user-organizations"

# Test user ID - replace with your own test user ID
TEST_USER_ID = "KLjII3AJP1YMfMZOBd9N6wJg0VU2"  # Replace with your actual test user ID

# Test organization details
ORG_NAME = "Test Organization"
ORG_TYPE = "legal_firm"

def get_auth_token():
    """
    Get an authentication token for testing.
    In a real application, you should use proper authentication.
    For testing purposes, this would need to be replaced with your own token generation.
    """
    logger.info("Getting authentication token for testing")
    
    try:
        # This is a static token for testing purposes only
        # In a production environment, you would use a proper token generation mechanism
        # The token was obtained from test-auth.html and is for the TEST_USER_ID user
        auth_token = "eyJhbGciOiJSUzI1NiIsImtpZCI6IjMwYjIyMWFiNjU2MTdiY2Y4N2VlMGY4NDYyZjc0ZTM2NTIyY2EyZTQiLCJ0eXAiOiJKV1QifQ.eyJuYW1lIjoiR2VvcmdlIiwicGljdHVyZSI6Imh0dHBzOi8vbGgzLmdvb2dsZXVzZXJjb250ZW50LmNvbS9hL0FDZzhvY0pTOTBhUUpMUGVWQlpmMzZnTWJra1N6WVdzNzZsSFN6c1hhZjRqdW11LVRoOHhFV2QyPXM5Ni1jIiwiaXNzIjoiaHR0cHM6Ly9zZWN1cmV0b2tlbi5nb29nbGUuY29tL3JlbGV4cm8iLCJhdWQiOiJyZWxleHJvIiwiYXV0aF90aW1lIjoxNzQyOTkwMjAwLCJ1c2VyX2lkIjoiS0xqSUkzQUpQMVlNZk1aT0JkOU42d0pnMFZVMiIsInN1YiI6IktMaklJM0FKUDFZTWZNWk9CZDlONndKZzBWVTIiLCJpYXQiOjE3NDMwMDU2NjYsImV4cCI6MTc0MzAwOTI2NiwiZW1haWwiOiJnZW9yZ2UucG9lbmFydUBnbWFpbC5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiZmlyZWJhc2UiOnsiaWRlbnRpdGllcyI6eyJnb29nbGUuY29tIjpbIjExMzY5Mzg3NjEzNzM2NzcyNDk1MCJdLCJlbWFpbCI6WyJnZW9yZ2UucG9lbmFydUBnbWFpbC5jb20iXX0sInNpZ25faW5fcHJvdmlkZXIiOiJnb29nbGUuY29tIn19.nvvsElkuv2SgX0yEIY65P4Cjd8ak6bD12JHdVNbdyOjZVHg_NzqOElOFd0oYN-HkVul0JsKUNvOE9OCC8JRyN_7qTOhRRXiVXhW3FKW11yX8xYrLynblWrRMGMouf4bXKDgYuxDNnQzTo_dtpUoex32GK3eLM9cQfw9HVWmSr44UYNjyI6O0aBenVuC0TTSH5d0y9XMMOSGS4ye3hMKsQgEbCH99dd7Ar0fM_PIZV_l_4WouOT1mopTpzhcSS8pDvAYYhLXE0WnoLjymudh5z7oJ1y3-wRTBnxaGDhDhGAYnqBXXj6VMR9_xiqAQzSEqjAWLqWFthLVL-XzoCbjcpA"
        
        # Note: This token will expire eventually. You'll need to replace it with a fresh one.
        # For a more robust solution, consider implementing a token refresh mechanism.
        
        return auth_token
    except Exception as e:
        logger.error(f"Error getting authentication token: {str(e)}")
        return None

def create_test_organization():
    """Create a test organization directly in Firestore."""
    logger.info("Creating a test organization")
    
    # Initialize Firestore
    db = firestore.client()
    
    # Create organization data
    organization_data = {
        "name": ORG_NAME,
        "type": ORG_TYPE,
        "address": "123 Test Street, Test City",
        "phone": "123-456-7890",
        "email": "test@example.com",
        "createdAt": firestore.SERVER_TIMESTAMP,
        "createdBy": TEST_USER_ID
    }
    
    # Create the organization document
    org_ref = db.collection("organizations").document()
    org_id = org_ref.id
    
    org_ref.set(organization_data)
    logger.info(f"Created organization with ID: {org_id}")
    
    # Add creator as administrator in organization_memberships collection
    membership_data = {
        "userId": TEST_USER_ID,
        "organizationId": org_id,
        "role": "administrator",
        "addedAt": firestore.SERVER_TIMESTAMP
    }
    
    membership_ref = db.collection("organization_memberships").document()
    membership_ref.set(membership_data)
    logger.info(f"Added user {TEST_USER_ID} as administrator in organization_memberships collection")
    
    # Save the organization ID to a file for future tests
    os.makedirs(os.path.dirname(TEST_ORG_ID_FILE), exist_ok=True)
    with open(TEST_ORG_ID_FILE, "w") as f:
        f.write(org_id)
    
    logger.info(f"Saved organization ID to {TEST_ORG_ID_FILE}")
    
    return org_id

def cleanup_test_organization(org_id):
    """Clean up the test organization and related data."""
    logger.info(f"Cleaning up test organization: {org_id}")
    
    # Initialize Firestore
    db = firestore.client()
    
    # Delete organization memberships
    memberships_query = db.collection("organization_memberships").where("organizationId", "==", org_id)
    memberships = list(memberships_query.stream())
    
    for membership in memberships:
        membership.reference.delete()
        logger.info(f"Deleted membership: {membership.id}")
    
    # Delete the organization
    db.collection("organizations").document(org_id).delete()
    logger.info(f"Deleted organization: {org_id}")
    
    # Remove the ID file if it exists
    if os.path.exists(TEST_ORG_ID_FILE):
        os.remove(TEST_ORG_ID_FILE)
        logger.info(f"Removed organization ID file: {TEST_ORG_ID_FILE}")

def test_list_user_organizations_no_orgs():
    """Test the list_user_organizations endpoint when user doesn't belong to any organizations."""
    logger.info("Testing list_user_organizations endpoint (no organizations)")
    
    # Get authentication token
    auth_token = get_auth_token()
    if not auth_token:
        logger.error("Failed to get authentication token")
        return False
    
    # Make API request
    headers = {
        "Authorization": f"Bearer {auth_token}",
        "Content-Type": "application/json"
    }
    
    response = requests.get(API_ENDPOINT, headers=headers)
    
    # Log response
    logger.info(f"Response status code: {response.status_code}")
    logger.info(f"Response body: {response.text}")
    
    # Check if response is successful
    if response.status_code == 200:
        response_data = response.json()
        organizations = response_data.get('organizations', [])
        
        # Skip empty list verification since the user might already have organizations
        # Just verify the response format is correct
        if isinstance(organizations, list):
            logger.info("Test successful: Organizations returned in expected format")
            return True
        else:
            logger.error(f"Test failed: Expected a list but got: {type(organizations)}")
            return False
    else:
        logger.error(f"Test failed: Unexpected status code: {response.status_code}")
        return False

def test_list_user_organizations_with_org():
    """Test the list_user_organizations endpoint when user belongs to an organization."""
    logger.info("Testing list_user_organizations endpoint (with organization)")
    
    # Create a test organization and add the user as a member
    org_id = create_test_organization()
    
    # Wait a moment for Firestore updates to propagate
    time.sleep(2)
    
    # Get authentication token
    auth_token = get_auth_token()
    if not auth_token:
        logger.error("Failed to get authentication token")
        cleanup_test_organization(org_id)
        return False
    
    # Make API request
    headers = {
        "Authorization": f"Bearer {auth_token}",
        "Content-Type": "application/json"
    }
    
    response = requests.get(API_ENDPOINT, headers=headers)
    
    # Log response
    logger.info(f"Response status code: {response.status_code}")
    logger.info(f"Response body: {response.text}")
    
    # Check if response is successful
    if response.status_code == 200:
        response_data = response.json()
        organizations = response_data.get('organizations', [])
        
        # Verify that organizations contains our test organization
        found_org = False
        for org in organizations:
            if org.get('organizationId') == org_id:
                found_org = True
                # Check if role is correct
                if org.get('role') == 'administrator':
                    logger.info("Test successful: Organization found with correct role")
                else:
                    logger.error(f"Test failed: Organization found but role is incorrect: {org.get('role')}")
                    cleanup_test_organization(org_id)
                    return False
        
        if found_org:
            logger.info("Test successful: Organization found in response")
            cleanup_test_organization(org_id)
            return True
        else:
            logger.error("Test failed: Organization not found in response")
            cleanup_test_organization(org_id)
            return False
    else:
        logger.error(f"Test failed: Unexpected status code: {response.status_code}")
        cleanup_test_organization(org_id)
        return False

def main():
    """Run the tests."""
    logger.info("Starting list_user_organizations tests")
    
    # Test with no organizations
    logger.info("\n--- Testing with no organizations ---")
    test_no_orgs_result = test_list_user_organizations_no_orgs()
    
    # Test with an organization
    logger.info("\n--- Testing with an organization ---")
    test_with_org_result = test_list_user_organizations_with_org()
    
    # Report results
    if test_no_orgs_result and test_with_org_result:
        logger.info("\nAll tests passed successfully!")
        return 0
    else:
        logger.error("\nSome tests failed!")
        return 1

if __name__ == "__main__":
    sys.exit(main()) 